// core/src/table/new-api/table-renderer.api.ts
export type TableScrollToNodeHandler = (nodeIndex: number, options?: { autoFocus?: boolean }) => void;
// ...
export namespace TableRenderPropsType {
	export interface ContextMenuBaseProps {
		closeHandler: () => void;
	}

	export interface ContextMenuProps<RowType = unknown> extends ContextMenuBaseProps {
		row: RowType;

		rowIndex: number;
	}

	export interface HeadContextMenuProps<ColumnType = unknown> extends ContextMenuBaseProps {
		column?: ColumnType;
	}

	export interface BaseProps extends Identifiable, Styleable, Ref<HTMLDivElement> {
		key?: Key;
	}

	export type HeadProps = Partial<TableTemplateProps.HeadRowProps>;

	export interface HeadRowProps extends BaseProps, Partial<TableTemplateProps.HeadRowProps> {}

	export interface RowGroupHeaderProps<RowType = unknown> extends Partial<TableTemplateProps.RowGroupHeaderProps> {
		row: RowType;
		rowIndex: number;
	}

	export interface HeadCellProps<ColumnType> extends BaseProps, Partial<TableTemplateProps.HeadCellProps> {
		column: ColumnType;
		resizableColumn?: ColumnType;
	}

	export interface HeadContentProps<ColumnType> extends BaseProps, Partial<TableTemplateProps.HeadCellProps> {
		column: ColumnType;
	}

	export interface BodyProps<RowType = unknown> extends Partial<TableTemplateProps.BodyProps> {
		data: RowType[];
		scrollToNode?(handler: TableScrollToNodeHandler): void;
	}

	export interface InfiniteScrollBodyProps<RowType = unknown> extends Omit<BodyProps<RowType>, "data"> {
		data: (RowType | undefined)[];

		infiniteScrollOptions: InfiniteScrollOptions;
	}

	export interface VirtualizedBodyProps<RowType = unknown> extends BodyProps<RowType> {
		virtualScrollOptions?: Partial<ReactVirtualizedProps> & {
			listRef?: RefCallback<ReactVirtualizedList>;
		};
		scrollToNode?(handler: TableScrollToNodeHandler): void;
	}

	export interface BodyRowProps<RowType = unknown> extends BaseProps, Partial<TableTemplateProps.BodyRowProps> {
		row: RowType;
		rowIndex: number;
		onRendered?: () => void;

		/**
		 * Useful in case of virtual scrolling to trigger a re-calculation of row height when content changes.
		 */
		measureRowHeight?: () => void;
		loading?: boolean;
	}

	export interface BodyCellProps<
		RowType = unknown,
		ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
	>
		extends BaseProps, Partial<TableTemplateProps.BodyCellProps> {
		row: RowType;
		rowIndex: number;
		column: ColumnType;
	}

	export interface BodyContentProps<
		RowType = unknown,
		ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
	> extends BaseProps {
		row: RowType;
		rowIndex: number;
		column: ColumnType;
	}

	export interface PlaceHolderBodyRowProps extends BaseProps {
		rowIndex: number;
		role?: string;
	}

	export interface PlaceHolderBodyCellProps<
		RowType = unknown,
		ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
	> extends BaseProps {
		rowIndex: number;
		column: ColumnType;
	}

	export interface PlaceHolderBodyContentProps<
		RowType = unknown,
		ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
	> extends BaseProps {
		rowIndex: number;
		column: ColumnType;
	}

	export interface FootProps extends BaseProps, Partial<TableTemplateProps.FootProps> {}

	export interface FootRowProps extends BaseProps, Partial<TableTemplateProps.FootRowProps> {}

	export interface FootCellProps<ColumnType> extends BaseProps, Partial<TableTemplateProps.FootCellProps> {
		column: ColumnType;
	}

	export interface FootContentProps<ColumnType> extends BaseProps {
		column: ColumnType;
	}

	export interface DragObject<RowType = unknown> {
		rowIndex: number;
		row: RowType;
	}

	export interface HoveredObject<RowType = unknown> {
		rowIndex: number;
		row: RowType;
	}

	export interface DropResult<RowType = unknown> {
		rowIndex: number;
		row: RowType;
	}

	export interface DndBodyRowProps<RowType = unknown> extends BodyRowProps<RowType> {
		/**
		 * @deprecated since 36.2.0 because drag and drop to the top area is now always allowed.
		 */
		shouldRenderTopDropTarget?: boolean;
		canDrag?: boolean;
	}

	export interface DragSourceProps<RowType = unknown> extends Container, Styleable, Identifiable {
		rowIndex: number;
		row: RowType;
		canDrag?: boolean;
		dragPreview?: ReactNode;

		/** Get the preview function from DragSourceLayer, useful for customizing the default DragLayer  */
		connectDragPreview?: (preview: ConnectDragPreview) => void;
	}

	export interface DropTargetProps<RowType = unknown> extends Container, Styleable, Identifiable {
		rowIndex: number;
		row: RowType;
	}

	export interface DragPreviewProps extends Container, Identifiable, Styleable {}

	export interface ColumnResizeHandlerProps extends Styleable {
		isRightResizeHandler?: boolean;
		isFirstColumnOfRightPin?: boolean;
		column: BaseColumnType;
	}
}